feat: Add code for FairSequence and FuelPurchase classes
This commit adds the FairSequence and FuelPurchase classes to the Day7 package. The FairSequence class contains a method to calculate the maximum sum of elements in a fair subsequence from an array. The FuelPurchase class contains a method to calculate the maximum overall volume of fuels that can be purchased given a limited amount of money. These classes provide solutions to the corresponding questions mentioned in the Readme.md file.